flask_server/requirements.txt
boto3
torch
streamlit
transformers
PyPDF4
flask
flask-cors
textblob
pandas
numpy
sentencepiece
sacremoses

